requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.4
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.2
django-import-export==3.3.7
django-phonenumber-field==7.3.0
django-phonenumbers==1.0.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
idna==3.6
inflection==0.5.1
Markdown==3.6
MarkupPy==1.14
# mysqlclient==2.2.4
numpy==1.26.4
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==24.1
pandas==2.2.1
phonenumbers==8.13.34
pillow==10.3.0
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
setuptools==74.0.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.4.4
tablib==3.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
xlrd==2.0.1
xlwt==1.3.0
